Move AABB, Matrix3, and Vector2 modules from lib to models
