+++ x-pack/plugins/security_solution/public/resolver/models/aabb.ts
