+++ x-pack/plugins/security_solution/public/resolver/models/matrix3.ts
