+++ x-pack/plugins/security_solution/public/resolver/models/vector2.ts
