+++ x-pack/plugins/security_solution/public/resolver/store/camera/reducer.ts
@@ -7,7 +7,7 @@
 import { Reducer } from 'redux';
 import { unitsPerNudge, nudgeAnimationDuration } from './scaling_constants';
 import { animatePanning } from './methods';
-import * as vector2 from '../../lib/vector2';
+import * as vector2 from '../../models/vector2';
 import * as selectors from './selectors';
 import { clamp } from '../../lib/math';
