+++ x-pack/plugins/security_solution/public/resolver/store/camera/zooming.test.ts
@@ -11,7 +11,7 @@ import { CameraState, AABB } from '../../types';
 import { viewableBoundingBox, inverseProjectionMatrix, scalingFactor } from './selectors';
 import { expectVectorsToBeClose } from './test_helpers';
 import { scaleToZoom } from './scale_to_zoom';
-import { applyMatrix3 } from '../../lib/vector2';
+import { applyMatrix3 } from '../../models/vector2';
 
 describe('zooming', () => {
   let store: Store<CameraState, CameraAction>;
