+++ x-pack/plugins/security_solution/public/resolver/view/edge_line.tsx
@@ -7,7 +7,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { FormattedMessage } from '@kbn/i18n/react';
-import { applyMatrix3, distance, angle } from '../lib/vector2';
+import { applyMatrix3, distance, angle } from '../models/vector2';
 import { Vector2, Matrix3, EdgeLineMetadata } from '../types';
 import { useResolverTheme, calculateResolverFontSize } from './assets';
