+++ x-pack/plugins/security_solution/public/resolver/view/use_camera.test.tsx
@@ -14,7 +14,7 @@ import { storeFactory } from '../store';
 import { Matrix3, ResolverStore, SideEffectSimulator } from '../types';
 import { ResolverEvent } from '../../../common/endpoint/types';
 import { SideEffectContext } from './side_effect_context';
-import { applyMatrix3 } from '../lib/vector2';
+import { applyMatrix3 } from '../models/vector2';
 import { sideEffectSimulator } from './side_effect_simulator';
 import { mockProcessEvent } from '../models/process_event_test_helpers';
 import { mock as mockResolverTree } from '../models/resolver_tree';
